+++ src/community/anushree-gupta.mdx
@@ -1,0 +1,5 @@
+---
+name: Anushree Gupta
+role: Researcher
+isCurrentContributor: false
+---
