+++ src/community/eishita-chhetri.mdx
@@ -1,0 +1,5 @@
+---
+name: Eishita Chhetri
+role: Illustrator
+isCurrentContributor: false
+---
