+++ src/community/harman-singh.mdx
@@ -1,0 +1,5 @@
+---
+name: Harman Singh
+role: Mobile Engineer
+isCurrentContributor: false
+---
