+++ src/community/j-mahalakshmi.mdx
@@ -1,0 +1,5 @@
+---
+name: J Mahalakshmi
+role: Data Scientist
+isCurrentContributor: true
+---
