+++ src/community/keshav-joshi.mdx
@@ -1,0 +1,5 @@
+---
+name: Keshav Joshi
+role: Data Scientist
+isCurrentContributor: false
+---
