+++ src/community/kruttika-nadig.mdx
@@ -1,0 +1,5 @@
+---
+name: Kruttika Nadig
+role: Data Scientist
+isCurrentContributor: false
+---
