+++ src/community/mohit-malik.mdx
@@ -1,0 +1,5 @@
+---
+name: Mohit Malik
+role: Full Stack Intern
+isCurrentContributor: false
+---
