+++ src/community/nilomee-jesrani.mdx
@@ -1,0 +1,6 @@
+---
+name: Nilomee Jesrani
+role: UX / Visual Design Contributor
+url: "http://nilomeejesrani.work/"
+isCurrentContributor: false
+---
