+++ src/community/rishabh-jain.mdx
@@ -1,0 +1,6 @@
+---
+name: Rishabh Jain
+role: Design Research
+url: "https://www.jaineric.com/"
+isCurrentContributor: false
+---
