+++ src/community/saakshita-prabhakar.mdx
@@ -1,0 +1,6 @@
+---
+name: Saakshita Prabhakar
+role: UX / Visual Design Contributor
+url: "https://saakshita.com"
+isCurrentContributor: true
+---
