+++ src/community/saumya-gupta.mdx
@@ -1,0 +1,5 @@
+---
+name: Saumya Gupta
+role: Journalist/Researcher
+isCurrentContributor: false
+---
