+++ src/community/sudeep-duggal.mdx
@@ -1,0 +1,6 @@
+---
+name: Sudeep Duggal
+role: Contributor: Data Science, Security, Accessibility
+url: "https://github.com/duggalsu"
+isCurrentContributor: true
+---
