+++ src/community/sunandan-chakraborty.mdx
@@ -1,0 +1,5 @@
+---
+name: Sunandan Chakraborty
+role: Past Advisor, (Assistant Professor, Information School, Indiana University)
+isCurrentContributor: false
+---
