+++ src/community/suraj-sharma.mdx
@@ -1,0 +1,6 @@
+---
+name: Suraj Sharma
+role: Full Stack Intern
+url: "http://evenzero.in"
+isCurrentContributor: false
+---
