+++ src/community/swair-shah.mdx
@@ -1,0 +1,5 @@
+---
+name: Swair Shah
+role: ML Engineer
+isCurrentContributor: true
+---
