+++ src/community/swastika-mohapatra.mdx
@@ -1,0 +1,6 @@
+---
+name: Swastika Mohapatra
+role: Data Science Intern
+url: "https://www.linkedin.com/in/swastika-mohapatra/"
+isCurrentContributor: false
+---
