+++ src/community/tarunima-prabhakar.mdx
@@ -1,0 +1,6 @@
+---
+name: Tarunima Prabhakar
+role: Research and Project Lead
+url: "https://tarunima.com"
+isCurrentContributor: true
+---
