+++ src/community/upasana-bhattacharjee.mdx
@@ -1,0 +1,5 @@
+---
+name: Upasana Bhattacharjee
+role: Qualitative Researcher
+isCurrentContributor: false
+---
