+++ src/community/yash-budhwar.mdx
@@ -1,0 +1,6 @@
+---
+name: Yash Budhwar
+role: Operations and Strategy Associate
+url: "http://www.linkedin.com/in/yash-budhwar-495290a5"
+isCurrentContributor: true
+---
